requirements.txt
pandas
fastapi
uvicorn
streamlit
fastai
torch
torchvision
pillow
requests
matplotlib
numpy
python-multipart
